feat: protocol version in window title
This makes it easier for users to understand what version they are playing on, as LiquidBounce is pretty much version independent except for the requirement to run on 1.21.1. Unfortunately I still get way too many questions about how to run LiquidBounce on e.g. 1.8.x, 1.16 etc. for some bizarre reason.