refactor: proxy validator
To improve accuracy when checking proxy health, we now query a Minecraft server instead. This ensures that the proxy supports Minecraft communication. There are many providers and proxies that only support HTTP/HTTPS communication, so this is the best way to check. This also has the side effect of giving us accurate ping information, which may be useful to implement in the GUI in the future.