fix(BlockShapeEvent): collision shape change on non-movement
With this fix, the collision shape is only changed when it is used for movement validation. This is useful to prevent problems with ShapeCache, e.g. "has a collision shape and on offset type, but is not marked as dynamicShape in it's properties" - this often caused people to be unable to join a server.