chore: MCEF 1.3.0
This update introduces a more robust resource manager - which will now make use of the new LiquidBounce Resource API:
https://api.liquidbounce.net/api/v3/resource/:name/:commit/:platform
https://api.liquidbounce.net/api/v3/resource/:name/:commit/:platform/checksum

This means from now on downloads are being served via Cloudflare R2 - which should be much faster and not rely on our infrastructure so much.

Error handling was slightly improved as well.