fix(ClickGUI/Null): use transition color instead of negated color (#5736)
Negating a color to make it visible on another color is a horrible choice when we want to make it look good. This function will either use black or white depending on the chosen color, while also fading between black and white when we reach a certain transition color range. This will allow rainbow color to have a smooth transition state.