fix(Clicker): flawed attack cooldown handling
The 'Consider Miss Cooldown' option was not working at all, because
1. When Auto Clicker was enabled, we would always have a cooldown of 0 due to a hacky patch by @mems01.
2. We did not mark the [KeyBinding] as pressed, when we clicked, so Minecraft automatically reset the attack cooldown back to zero.

Also:
- Cooldown is now renamed to Item Cooldown
- Consider Miss Cooldown is now renamed to Attack Cooldown (To follow the Minecraft logic).

Attack Cooldown integration is now tightly built into Clicker, which allows modules that rely on the [attackKey] to use the function as well. It's also enabled by default now.

This WILL decrease CPS on e.g. Auto Clicker when missing!